# api_gateway/docs/swagger.yaml
basePath: /
definitions:
...
  dto.Response:
    properties:
      data:
        x-order: "1"
      message:
        type: string
        x-order: "0"
    type: object
...
  dto.SwaggerResponseRegister:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/models.User'
        x-order: "1"
      message:
        type: string
        x-order: "0"
    type: object
...
  dto.UserLogin:
    properties:
      email:
        type: string
        x-order: "0"
      password:
        type: string
        x-order: "1"
    required:
    - email
    - password
    type: object
  dto.UserRegister:
    properties:
      birth_date:
        type: string
        x-order: "4"
      email:
        type: string
        x-order: "2"
      first_name:
        type: string
        x-order: "0"
      last_name:
        type: string
        x-order: "1"
      password:
        type: string
        x-order: "3"
    required:
    - birth_date
    - email
    - first_name
    - password
    type: object
...
  models.User:
    properties:
      birth_date:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
...
  utils.ErrResponse:
    properties:
      detail:
        type: string
        x-order: "2"
      status:
        type: integer
        x-order: "0"
      type:
        type: string
        x-order: "1"
    type: object
host: jeki-x72poumuyq-et.a.run.app
info:
  contact:
    email: daniel.rahmanto@gmail.com
    name: Contact the developer
  description: Food delivery app built with microservices that integrates customer,
    driver, and restaurant.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Jeki
  version: "1.0"
paths:
...
  /users/login:
    post:
      consumes:
      - application/json
      description: Log users in and embeds a JWt-Auth in cookie
      parameters:
      - description: User login details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Log in
      tags:
      - all user
  /users/logout:
    get:
      consumes:
      - application/json
      description: Logout the currently authenticated user and clears the authorization
        cookie
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Logout the user
      tags:
      - all user
...
  /users/register/admin:
    post:
      consumes:
      - application/json
      description: Register a new user with the role 'Admin'
      parameters:
      - description: Admin registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserRegister'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SwaggerResponseRegister'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Register a new admin
      tags:
      - merchant
  /users/register/driver:
    post:
      consumes:
      - application/json
      description: Register a new user with the role 'Driver'
      parameters:
      - description: Driver registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserRegister'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SwaggerResponseRegister'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Register a new driver
      tags:
      - driver
  /users/register/user:
    post:
      consumes:
      - application/json
      description: Register a new user with the role 'User'
      parameters:
      - description: User registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserRegister'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SwaggerResponseRegister'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Register a new user
      tags:
      - customer
  /users/verify/{userid}/{token}:
    get:
      consumes:
      - application/json
      description: Verify the user registration using unique token sent to the registered
        email
      parameters:
      - description: User ID
        in: path
        name: userid
        required: true
        type: integer
      - description: Verification token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Verify user credentials
      tags:
      - all user
swagger: "2.0"